___JmolVersion="13.1.16_dev_2013.05.28e"
Former-commit-id: 25bf0d56936227538e791f0946cfffcd29e3cc26

+++ Jmol/src/org/jmol/adapter/readers/pymol/PyMOLReader.java
@@ -1460,7 +1460,7 @@ public class PyMOLReader extends PdbReader implements PymolAtomReader {
   }
 
   public float getVDW(int iAtom) {
-    return (int) atoms[iAtom].radius;
+    return atoms[iAtom].radius;
   }
 
   public int getSequenceNumber(int iAtom) {
